+++ Sources/Resources/Orgs/OrgsClient.swift
@@ -31,6 +31,18 @@ public final class OrgsClient: Sendable {
         )
     }
 
+    /// Retrieves private organization data for the caller's organization. | authz: min_org_role=operator | () -> (OrgPrivateData1)
+    ///
+    /// - Parameter requestOptions: Additional options for configuring the request, such as custom headers or timeout settings.
+    public func getOrgPrivateDataV1(requestOptions: RequestOptions? = nil) async throws -> OrgPrivateData1 {
+        return try await httpClient.performRequest(
+            method: .get,
+            path: "/orgs/org_private_data/v1",
+            requestOptions: requestOptions,
+            responseType: OrgPrivateData1.self
+        )
+    }
+
     /// Retrieves public organization data for the caller's organization. | () -> (OrgPublicData1)
     ///
     /// - Parameter requestOptions: Additional options for configuring the request, such as custom headers or timeout settings.